### 2023-04-24 #589 @derek-corcoran-barrios
AICcPermanova: Model Selection of PERMANOVA Models Using AICc
Submitting Author Name: Derek Corcoran
Submitting Author Github Handle: @derek-corcoran-barrios
Repository:   https://github.com/Sustainscapes/AICcPerm
Version submitted: 0.0.3
Submission type: Standard

Version accepted: TBD

---

```
Package: AICcPermanova
Title: Model Selection of PERMANOVA Models Using AICc
Version: 0.0.3
    person(given = "Derek", family ="Corcoran",
    email = "derek.corcoran.barrios@gmail.com", role = c("aut", "cre"))
Description: Provides tools for model selection and model averaging of PerMANOVA
    models using Akaike Information Criterion corrected for small sample sizes
    (AICc) and Information Theoretic criteria principles. The package is built
    around the PERMANOVA analysis from the 'vegan' package and provides a
    streamlined workflow for generating and comparing models, obtaining model
    weights, and summarizing results using model averaging approaches.  The
    methods implemented in this package are based on the practical information-
    theoretic approach described by Burnham, K. P. and Anderson, D. R. (2002)
    (<doi:10.1007/b97636>).
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    broom,
    car,
    doParallel,
    foreach,
    furrr,
    future,
    parallel,
    vegan
    https://sustainscapes.github.io/AICcPerm/
    covr,
```

The package I build is built to generate a clear workflow for community analysis, specifically for model selection in Permanova,

Community ecologist in general, I see it as very useful for microbiologists working with dna barcoding taxonomy, botanist that do plots for community analysis, etc.

-
This package is based on the vegan package, so that package has the analysis made here, however it does not have the model selection capabilities. MuMIn does model selection but it cant deal with the PERMANOVA

https://github.com/ropensci/software-review/issues/585

The manuscript is probably 3 weeks away from finished, but it is almost there
