### 2023-05-01 #589 @derek-corcoran-barrios
Thanks Mauro, One more question should I only put one category, so
r"egression and supervised learning" or should I add "workflow automation"
ass well?

Cheers

Derek

On Sun, Apr 30, 2023 at 12:43 AM Mauro Lepore ***@***.***>
wrote:

> Dear @derek-corcoran-barrios <https://github.com/derek-corcoran-barrios>,
>
> Sorry for misleading you.
>
> Two of our stats experts think this should be a stats submission and the
> category you originally suggested "regression and supervised learning"
> seems like a good fit.
>
> Unfortunately the template is different so I'll close this issue now and
> ask you to please open a new issue choosing "Submit statistical software
> for review"
> <https://github.com/ropensci/software-review/issues/new?assignees=&labels=&template=F-submit-statistical-software-for-review.md>
> .
>
> Also FYI my EiC rotation is coming to an end soon so likely you'll
> continue with @maelle <https://github.com/maelle>.
>
> All the best in the process.
>
