### 2023-05-01 #589 @maurolepore
I think it'll be enough with "regression and supervised learning". I'll point @maelle to the editor's thread where that category was suggested.
